Price-Monitor/requirements.txt
pandas
requests
beautifulsoup4
lxml
gspread
google-auth
dotenv